iommu/arm-smmu-v3: Add cd_table user counter
This can be used to check if some CD Table Entry is still live when the
driver decides to tear down the cd table. Since arm_smmu_free_cd_tables
is called in arm_smmu_release_device that cannot be re-entered, warn if
user counter is unbalanced but carry on the release of the CD table.

Change-Id: I0eb39ab0784b67eb418e9e6fbb6f6708f1c98544
Suggested-by: Jason Gunthorpe <jgg@nvidia.com>
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>