iommu/arm-smmu-v3: Add a robust algorithm for arm_smmu_write_strtab_ent
The current arm_smmu_write_strtab_ent is not perfectly following the spec,
but taking a shortcut by staging an STE to abort/bypass mode and limiting
its supported scenarios.

Chapter "3.21.3.1 Configuration structure update procedure" in the SMMUv3
spec provides a flow to updating an STE or a CD.

Follow the spec and write an algorithm for STE updates.

Change-Id: Ib11637cc5d4f2519fb00c1290d6cfd06d898d818
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>