iommu/arm-smmu-v3: Support IOMMU_DOMAIN_NESTED type of allocations
Add domain allocation support for the IOMMU_DOMAIN_NESTED type that is
used to represent a user space stage-1 context description table. Save
all the stage-1 STE fields in master->cd_table to take part in the STE
calculation.

Also store the user space Stream ID in an xarray and then link it with
the host-level Stream ID, which will be later used to replace the user
SID in a user space ATC_INV command to invalidate the physical cache.

Co-developed-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>