iommufd: Add IOMMUFD_OBJ_VIOMMU and IOMMUFD_CMD_VIOMMU_ALLOC
On a multi-IOMMU system, the VMM has to be aware of the different IOMMU
instances, each of which maintains its own driver speicific info. For
example, NVIDIA's VCMDQ for SMMUv3 requires a pair of configurations of
physical Steam ID and virtual Stream ID. Each SMMU instance per VM has
to maintain its own Stream ID list.

Introduce a new IOMMUFD_OBJ_VIOMMU to represent an IOMMU instance that's
mentioned above. And add a new ioctl for user space to allocate it.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>