iommufd/viommu: Add IOMMU_VDEVICE_ALLOC ioctl
Introduce a new ioctl to allocate a vDEVICE object. Since a vDEVICE object
is a connection of an iommufd_iommufd object and an iommufd_viommu object,
require both as the ioctl inputs and take refcounts in the ioctl handler.

Add to the vIOMMU object a "vdevs" xarray, indexed by a per-vIOMMU virtual
device ID, which can be:
 - Virtual StreamID on a nested ARM SMMUv3, an index to a Stream Table
 - Virtual DeviceID on a nested AMD IOMMU, an index to a Device Table
 - Virtual ID on a nested Intel VT-D IOMMU, an index to a Context Table

Add a rw_semaphore protection around the vdevs list. A function that wants
to access the list must grab the lock with a read or write permission.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>